@@ -3,13 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@chakra-ui/icons": "^1.0.1",
-    "@chakra-ui/react": "^1.8.8",
-    "@emotion/react": "^11.0.0",
-    "@emotion/styled": "^11.0.0",
-    "framer-motion": "^4.1.17",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "@chakra-ui/react": "^2.7.1",
+    "@chakra-ui/icons": "^2.0.19",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "framer-motion": "^6.5.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^3.11.0",
     "react-scripts": "^5.0.1"
   },
   "homepage": "https://fede.dm/FIUBA-Repos",
@@ -34,7 +35,6 @@
     ]
   },
   "devDependencies": {
-    "fireworks-js": "^1.3.5",
-    "use-resize-observer": "^7.0.0"
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11"
   }
 }
