@@ -1,11 +1,13 @@
 import { ChakraProvider } from "@chakra-ui/react";
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import MainApp from "./components/MainApp";
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = ReactDOM.createRoot(container);
+
+root.render(
   <ChakraProvider>
     <MainApp />
-  </ChakraProvider>,
-  document.getElementById("root")
+  </ChakraProvider>
 );
